pyproject.toml
[project]
name = "stable-diffusion-in-pytorch"
version = "0.1.0"
description = "Stable Diffusion implementation in PyTorch"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "pytorch",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["stable_diffusion*"]
